Move pdsh utility functions/classes to a seperate module
